mejoras models
